+++ packages/create-sitecore-jss/src/templates/nextjs-multisite/scripts/config/plugins/multisite.ts
@@ -14,29 +14,15 @@ class MultisitePlugin implements ConfigPlugin {
 
   async exec(config: JssConfig) {
     let sites: SiteInfo[] = [];
-
-    const endpoint = config.sitecoreEdgeContextId ? config.sitecoreEdgeUrl : config.graphQLEndpoint;
-
-    if (!endpoint || process.env.SITECORE) {
-      console.warn(
-        chalk.yellow(
-          `Skipping site information fetch (${
-            !endpoint ? 'missing GraphQL endpoint' : 'building on XM Cloud'
-          })`
-        )
-      );
-    } else {
-      console.log(`Fetching site information from ${endpoint}`);
-
-      try {
-        const siteInfoService = new GraphQLSiteInfoService({
-          clientFactory: createGraphQLClientFactory(config),
-        });
-        sites = await siteInfoService.fetchSiteInfo();
-      } catch (error) {
-        console.error(chalk.red('Error fetching site information'));
-        console.error(error);
-      }
+    console.log('Fetching site information');
+    try {
+      const siteInfoService = new GraphQLSiteInfoService({
+        clientFactory: createGraphQLClientFactory(config),
+      });
+      sites = await siteInfoService.fetchSiteInfo();
+    } catch (error) {
+      console.error(chalk.red('Error fetching site information'));
+      console.error(error);
     }
 
     return Object.assign({}, config, {