+++ packages/create-sitecore-jss/src/templates/nextjs-xmcloud/src/Scripts.tsx
@@ -1,0 +1,14 @@
+// The BYOC bundle imports external (BYOC) components into the app and makes sure they are ready to be used
+import BYOC from 'src/byoc';
+import CdpPageView from 'components/CdpPageView';
+
+const Scripts = (): JSX.Element => {
+  return (
+    <>
+      <BYOC />
+      <CdpPageView />
+    </>
+  );
+};
+
+export default Scripts;