+++ packages/create-sitecore-jss/src/templates/react/scripts/generate-config.js
@@ -9,17 +9,21 @@ const { jssConfigFactory } = require('./config');
 /**
  * REACT_APP is the default prefix for environment variables that will be made available to the app.
  */
+// REACT_APP_JSS_APP_NAME env variable has been deprecated since v.21.6, REACT_APP_SITE_NAME should be used instead
 const defaultConfig = {
   sitecoreApiKey: process.env[`${constantCase('reactAppSitecoreApiKey')}`],
   sitecoreApiHost: process.env[`${constantCase('reactAppSitecoreApiHost')}`],
-  jssAppName: process.env[`${constantCase('reactAppJssAppName')}`],
+  sitecoreSiteName:
+    process.env[`${constantCase('reactAppSitecoreSiteName')}`] ||
+    process.env[`${constantCase('reactAppJssAppName')}`],
   graphQLEndpointPath: process.env[`${constantCase('reactAppGraphQLEndpointPath')}`],
   defaultLanguage: process.env[`${constantCase('reactAppDefaultLanguage')}`],
   graphQLEndpoint: process.env[`${constantCase('reactAppGraphQLEndpoint')}`],
-  layoutServiceConfigurationName: process.env[`${constantCase('reactAppLayoutServiceConfigurationName')}`],
+  layoutServiceConfigurationName:
+    process.env[`${constantCase('reactAppLayoutServiceConfigurationName')}`],
 };
 
-generateConfig()
+generateConfig();
 
 /**
  * Generate config
@@ -30,7 +34,7 @@ generateConfig()
  */
 function generateConfig() {
   try {
-    config = jssConfigFactory.create(defaultConfig)
+    config = jssConfigFactory.create(defaultConfig);
   } catch (error) {
     console.error('Error generating config');
     console.error(error);
@@ -38,7 +42,7 @@ function generateConfig() {
   }
 
   writeConfig(config);
-};
+}
 
 function writeConfig(config) {
   let configText = `/* eslint-disable */
@@ -47,7 +51,7 @@ function writeConfig(config) {
 const config = {};\n`;
 
   // Set base configuration values, allowing override with environment variables
-  Object.keys(config).forEach((prop) => {
+  Object.keys(config).forEach(prop => {
     configText += `config.${prop} = process.env.REACT_APP_${constantCase(prop)} || "${
       config[prop]
     }",\n`;