+++ packages/create-sitecore-jss/src/templates/vue/scripts/disconnected-mode-proxy.js
@@ -21,7 +21,7 @@ const touchToReloadFilePath = 'src/temp/config.js';
 
 const proxyOptions = {
   appRoot: path.join(__dirname, '..'),
-  appName: config.jssAppName,
+  appName: config.sitecoreSiteName,
   watchPaths: ['./data'],
   language: config.defaultLanguage,
   port: process.env.PROXY_PORT || 3042,