+++ packages/sitecore-jss-nextjs/src/context/index.ts
@@ -1,0 +1,1 @@
+export { Context, ContextConfig, SDK } from './context';