+++ packages/sitecore-jss-nextjs/src/middleware/personalize-middleware.ts
@@ -3,22 +3,22 @@ import {
   GraphQLPersonalizeService,
   GraphQLPersonalizeServiceConfig,
   getPersonalizedRewrite,
-  PosResolver,
+  PersonalizeInfo,
 } from '@sitecore-jss/sitecore-jss/personalize';
-import { SiteInfo } from '@sitecore-jss/sitecore-jss/site';
 import { debug } from '@sitecore-jss/sitecore-jss';
 import { MiddlewareBase, MiddlewareBaseConfig } from './middleware';
-import { initServer, EngageServer } from '@sitecore/engage';
+import { init, personalize } from '@sitecore-cloudsdk/personalize/server';
 
 export type CdpServiceConfig = {
   /**
-   * Your Sitecore CDP API endpoint
+   * Your Sitecore Edge Platform endpoint
+   * Default is https://edge-platform.sitecorecloud.io
    */
-  endpoint: string;
+  sitecoreEdgeUrl?: string;
   /**
-   * The client key to use for authentication
+   * Your unified Sitecore Edge Context Id
    */
-  clientKey: string;
+  sitecoreEdgeContextId: string;
   /**
    * The Sitecore CDP channel to use for events. Uses 'WEB' by default.
    */
@@ -42,13 +42,6 @@ export type PersonalizeMiddlewareConfig = MiddlewareBaseConfig & {
    * Configuration for your Sitecore CDP endpoint
    */
   cdpConfig: CdpServiceConfig;
-  /**
-   * function to resolve point of sale for a site
-   * @param {Siteinfo} site to get info from
-   * @param {string} language to get info for
-   * @returns point of sale value for site/language combination
-   */
-  getPointOfSale?: (site: SiteInfo, language: string) => string;
 };
 
 /**
@@ -101,22 +94,55 @@ export class PersonalizeMiddleware extends MiddlewareBase {
     };
   }
 
-  protected initializeEngageServer(
-    hostName: string,
-    site: SiteInfo,
-    language: string
-  ): EngageServer {
-    const engageServer = initServer({
-      clientKey: this.config.cdpConfig.clientKey,
-      targetURL: this.config.cdpConfig.endpoint,
-      pointOfSale: this.config.getPointOfSale
-        ? this.config.getPointOfSale(site, language)
-        : PosResolver.resolve(site, language),
-      cookieDomain: hostName,
-      forceServerCookieMode: true,
-    });
+  protected async initPersonalizeServer({
+    hostname,
+    siteName,
+    request,
+    response,
+  }: {
+    hostname: string;
+    siteName: string;
+    request: NextRequest;
+    response: NextResponse;
+  }): Promise<void> {
+    await init(
+      {
+        sitecoreEdgeUrl: this.config.cdpConfig.sitecoreEdgeUrl,
+        sitecoreEdgeContextId: this.config.cdpConfig.sitecoreEdgeContextId,
+        siteName,
+        cookieDomain: hostname,
+        enableServerCookie: true,
+      },
+      request,
+      response
+    );
+  }
 
-    return engageServer;
+  protected async personalize(
+    {
+      params,
+      personalizeInfo,
+      language,
+      timeout,
+    }: {
+      personalizeInfo: PersonalizeInfo;
+      params: ExperienceParams;
+      language: string;
+      timeout?: number;
+    },
+    request: NextRequest
+  ) {
+    const personalizationData = {
+      channel: this.config.cdpConfig.channel || 'WEB',
+      currency: this.config.cdpConfig.currency ?? 'USD',
+      friendlyId: personalizeInfo.contentId,
+      params,
+      language,
+    };
+
+    return (await personalize(personalizationData, request, timeout)) as {
+      variantId: string;
+    };
   }
 
   protected getExperienceParams(req: NextRequest): ExperienceParams {
@@ -194,36 +220,27 @@ export class PersonalizeMiddleware extends MiddlewareBase {
       return response;
     }
 
-    const engageServer = this.initializeEngageServer(hostname, site, language);
+    await this.initPersonalizeServer({
+      hostname,
+      siteName: site.name,
+      request: req,
+      response,
+    });
 
-    // creates the browser ID cookie on the server side
-    // and includes the cookie in the response header
-    try {
-      await engageServer.handleCookie(req, response, timeout);
-    } catch (error) {
-      debug.personalize('skipped (browser id generation failed)');
-      throw error;
-    }
     const params = this.getExperienceParams(req);
 
-    debug.personalize('executing experience for %s %s %o', personalizeInfo.contentId, params);
-
-    const personalizationData = {
-      channel: this.config.cdpConfig.channel || 'WEB',
-      currency: this.config.cdpConfig.currency ?? 'USA',
-      friendlyId: personalizeInfo.contentId,
-      params,
-      language,
-    };
+    debug.personalize('executing experience for %s %o', personalizeInfo.contentId, params);
 
     let variantId;
 
-    // Execute targeted experience in CDP
+    // Execute targeted experience in Personalize SDK
     // eslint-disable-next-line no-useless-catch
     try {
-      variantId = ((await engageServer.personalize(personalizationData, req, timeout)) as {
-        variantId: string;
-      }).variantId;
+      const personalization = await this.personalize(
+        { personalizeInfo, params, language, timeout },
+        req
+      );
+      variantId = personalization.variantId;
     } catch (error) {
       throw error;
     }
@@ -243,28 +260,11 @@ export class PersonalizeMiddleware extends MiddlewareBase {
 
     // Rewrite to persononalized path
     const rewritePath = getPersonalizedRewrite(basePath, { variantId });
-    // Note an absolute URL is required: https://nextjs.org/docs/messages/middleware-relative-urls
-    const rewriteUrl = req.nextUrl.clone();
-
-    // Preserve cookies from previous response
-    const cookies = response.cookies.getAll();
-
-    rewriteUrl.pathname = rewritePath;
-    response = NextResponse.rewrite(rewriteUrl, response);
+    response = this.rewrite(rewritePath, req, response);
 
     // Disable preflight caching to force revalidation on client-side navigation (personalization may be influenced)
     // See https://github.com/vercel/next.js/issues/32727
     response.headers.set('x-middleware-cache', 'no-cache');
-    // Share rewrite path with following executed middleware
-    response.headers.set('x-sc-rewrite', rewritePath);
-    // Share site name with the following executed middlewares
-    response.cookies.set(this.SITE_SYMBOL, site.name);
-
-    // Restore cookies from previous response since
-    // browserId cookie gets omitted after rewrite
-    cookies.forEach((cookie) => {
-      response.cookies.set(cookie);
-    });
 
     debug.personalize('personalize middleware end in %dms: %o', Date.now() - startTimestamp, {
       rewritePath,