+++ packages/sitecore-jss/src/site/graphql-siteinfo-service.test.ts
@@ -1,10 +1,17 @@
+/* eslint-disable no-unused-expressions */
 /* eslint-disable @typescript-eslint/no-unused-vars */
-import { expect } from 'chai';
+import { expect, spy, use } from 'chai';
+import spies from 'chai-spies';
 import nock from 'nock';
 import { GraphQLSiteInfoService, GraphQLSiteInfoResult } from './graphql-siteinfo-service';
 import { GraphQLRequestClient, PageInfo } from '../graphql';
+import debugApi from 'debug';
+import debug from '../debug';
+
+use(spies);
 
 describe('GraphQLSiteInfoService', () => {
+  let debugNamespaces: string;
   const endpoint = 'http://site';
   const apiKey = 'some-api-key';
 
@@ -12,17 +19,14 @@ describe('GraphQLSiteInfoService', () => {
     name,
     hostName,
     language,
-    pointOfSale = '',
   }: {
     name: string;
     hostName: string;
     language: string;
-    pointOfSale: string;
   }): GraphQLSiteInfoResult => ({
     name: { value: name },
     hostName: { value: hostName },
     language: { value: language },
-    pointOfSale: { value: pointOfSale },
   });
 
   const nonEmptyResponse = ({
@@ -45,7 +49,6 @@ describe('GraphQLSiteInfoService', () => {
               name: `site ${start + n}`,
               hostName: 'restricted.gov',
               language: 'en',
-              pointOfSale: 'en=en-pos',
             })
           ),
           ...sites,
@@ -63,8 +66,23 @@ describe('GraphQLSiteInfoService', () => {
     },
   };
 
+  before(() => {
+    debugNamespaces = debugApi.disable();
+    debugApi.enable(debug.multisite.namespace);
+  });
+
+  beforeEach(() => {
+    spy.on(debug.multisite, 'log', () => true);
+  });
+
   afterEach(() => {
     nock.cleanAll();
+    spy.restore(debug.multisite);
+    delete process.env.SITECORE;
+  });
+
+  after(() => {
+    debugApi.enable(debugNamespaces);
   });
 
   const mockSiteInfoRequest = (response: { [key: string]: unknown }) => {
@@ -81,7 +99,6 @@ describe('GraphQLSiteInfoService', () => {
             name: 'public 0',
             hostName: 'pr.showercurtains.org',
             language: '',
-            pointOfSale: '',
           }),
         ],
       })
@@ -93,15 +110,11 @@ describe('GraphQLSiteInfoService', () => {
         name: 'site 0',
         hostName: 'restricted.gov',
         language: 'en',
-        pointOfSale: {
-          en: 'en-pos',
-        },
       },
       {
         name: 'public 0',
         hostName: 'pr.showercurtains.org',
         language: '',
-        pointOfSale: undefined,
       },
     ]);
   });
@@ -114,7 +127,6 @@ describe('GraphQLSiteInfoService', () => {
             name: 'public 0',
             hostName: 'pr.showercurtains.org',
             language: '',
-            pointOfSale: '',
           }),
         ],
       })
@@ -130,15 +142,11 @@ describe('GraphQLSiteInfoService', () => {
         name: 'site 0',
         hostName: 'restricted.gov',
         language: 'en',
-        pointOfSale: {
-          en: 'en-pos',
-        },
       },
       {
         name: 'public 0',
         hostName: 'pr.showercurtains.org',
         language: '',
-        pointOfSale: undefined,
       },
     ]);
   });
@@ -159,49 +167,31 @@ describe('GraphQLSiteInfoService', () => {
         name: 'site 0',
         hostName: 'restricted.gov',
         language: 'en',
-        pointOfSale: {
-          en: 'en-pos',
-        },
       },
       {
         name: 'site 1',
         hostName: 'restricted.gov',
         language: 'en',
-        pointOfSale: {
-          en: 'en-pos',
-        },
       },
       {
         name: 'site 2',
         hostName: 'restricted.gov',
         language: 'en',
-        pointOfSale: {
-          en: 'en-pos',
-        },
       },
       {
         name: 'site 3',
         hostName: 'restricted.gov',
         language: 'en',
-        pointOfSale: {
-          en: 'en-pos',
-        },
       },
       {
         name: 'site 4',
         hostName: 'restricted.gov',
         language: 'en',
-        pointOfSale: {
-          en: 'en-pos',
-        },
       },
       {
         name: 'site 5',
         hostName: 'restricted.gov',
         language: 'en',
-        pointOfSale: {
-          en: 'en-pos',
-        },
       },
     ]);
   });
@@ -235,7 +225,6 @@ describe('GraphQLSiteInfoService', () => {
             name: 'public 0',
             hostName: 'pr.showercurtains.org',
             language: '',
-            pointOfSale: '',
           }),
         ],
       })
@@ -251,15 +240,11 @@ describe('GraphQLSiteInfoService', () => {
         name: 'site 0',
         hostName: 'restricted.gov',
         language: 'en',
-        pointOfSale: {
-          en: 'en-pos',
-        },
       },
       {
         name: 'public 0',
         hostName: 'pr.showercurtains.org',
         language: '',
-        pointOfSale: undefined,
       },
     ]);
     nock.cleanAll();
@@ -268,5 +253,17 @@ describe('GraphQLSiteInfoService', () => {
       .reply(200, emptyResponse);
     const resultCached = await service.fetchSiteInfo();
     expect(resultCached).to.deep.equal([]);
+  });
+
+  it('should skip on XM Cloud', async () => {
+    process.env.SITECORE = 'true';
+    nock(endpoint)
+      .post('/')
+      .reply(200, emptyResponse);
+    const service = new GraphQLSiteInfoService({ apiKey: apiKey, endpoint: endpoint });
+    const result = await service.fetchSiteInfo();
+    expect(result).to.deep.equal([]);
+    expect(debug.multisite.log, 'log debug message').to.be.called.once;
+    expect(nock.isDone(), 'skip request').to.be.false;
   });
 });